% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd.R
\docType{data}
\name{cd}
\alias{cd}
\title{CD Data Set}
\format{A data frame with 11 variables: \code{pk}, \code{id}, \code{artist}, 
\code{title}, \code{category}, \code{genre}, \code{cdextra}, \code{year}, 
\code{track_number}, \code{song_name} 

This data set includes 9763 CDs randomly extracted from freeDB.
It is appropriate for performing various types of record linkage
and can be assessed by standard record linkage methods.}
\usage{
cd
}
\description{
This data set includes 9763 CDs randomly extracted from freeDB.
}
\examples{
head(cd)
dim(cd)

}
\keyword{datasets}
