% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_gold.R
\docType{data}
\name{cd_gold}
\alias{cd_gold}
\title{CD Gold}
\format{A data frame with 2 variables: \code{disc1_id}, \code{disc2_id}

This data set includes the matched record pairs based on disc ID from the cd data set.
The data set can be used to evaluate the performance of record linkage methods performed on the cd data set.}
\usage{
cd_gold
}
\description{
This data set includes the matched record pairs based on disc ID.
}
\examples{
head(cd_gold)
dim(cd_gold)

}
\keyword{datasets}
