% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fodors.R
\docType{data}
\name{fodors}
\alias{fodors}
\title{Fodors Restaurants}
\format{A data frame with 5 variables: \code{name}, \code{address}, \code{city}, \code{phone}, \code{type}.

The Fodors dataset contains 533 records.
It is appropriate for performing various types of record linkage
and can be assessed by standard record linkage methods.}
\usage{
fodors
}
\description{
This dataset has information on 533 restaurants from the Fodors database
}
\examples{
head(fodors)
dim(fodors)

}
\keyword{datasets}
