% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\docType{data}
\name{restaurant}
\alias{restaurant}
\title{Restaurants data set}
\format{A data frame with 16 variables: \code{id}, \code{name}, \code{address}, \code{city}, \code{phone}, \code{type}, \code{class}

This data set includes 533 restaurants from the Fodors database and 331 records from the Zagat database. 
It is appropriate for performing various types of record linkage
and can be assessed by standard record linkage methods.}
\usage{
restaurant
}
\description{
This provides a record linkage data set with information about different restaurants.
}
\examples{
head(restaurant)
dim(restaurant)

}
\keyword{datasets}
