% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restaurant.R
\docType{data}
\name{restaurant_gold}
\alias{restaurant_gold}
\title{Restaurants Gold}
\format{A data frame with 3 variables: \code{class}, \code{id1}, \code{id2}

This data set includes the matched record pairs based on ID from the Zagats and Fodors data sets.
This data set can be used to evaluate the performance of record linkage methods performed on the Zagats and Fodors data sets.}
\usage{
restaurant_gold
}
\description{
This data set includes the matched record pairs based on ID.
}
\examples{
head(restaurant_gold)
dim(restaurant_gold)

}
\keyword{datasets}
