% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zagats.R
\docType{data}
\name{zagats}
\alias{zagats}
\title{Zagats Restaurants}
\format{A data frame with 5 variables: \code{name}, \code{address}, \code{city}, \code{phone}, \code{type}.

The Zagats dataset contains 331 records.
It is appropriate for performing various types of record linkage
and can be assessed by standard record linkage methods.}
\usage{
zagats
}
\description{
This dataset has information on 331 restaurants from the Zagats database
}
\examples{
head(zagats)
dim(zagats)


}
\keyword{datasets}
